haskell: modify & document System.PinkTrace module
Prepend function names with "pink" so that the module can be imported
without the qualified statement.
Define stronger type for Index. Before it was just a type synonym to
Int. Instead define it like:
  data Index = First | Second | Third | Fourth | Fifth
    deriving (Eq,Ord,Enum,Read,Show,Bounded)
which is stronger and less error-prone.
Remove the Length type synonym. It's not a common type.